Investigating quantization, scaling and functional correctness
Additions to the hardware module to correctly do the fixed point arithmetic
at full multiplication and addition bit width requirements. There is still no
logic to correctly do the rounding as this is still a work in progress but can
at least slice and truncate to a specifc width

Continue to ad fixed point capability to the python libraries

Additionally working on a new way to test the ospfb with impulse inputs that
provide structured output to help identify functional correctness and get a
better idea for scaling correct implementation

+++ python/apps/test_rfdc.py
@@ -17,9 +17,12 @@ if __name__=="__main__":
   nsamps = 256
 
   xt = np.zeros(nsamps, dtype=np.complex64)
-  xn_re = np.zeros(nsamps)
-  xn_im = np.zeros(nsamps)
+  xn_re = np.zeros(nsamps, dtype=np.int16)
+  xn_im = np.zeros(nsamps, dtype=np.int16)
   for i in range(0, nsamps):
     s = src.genSample()
     xt[i] = s
     xn_im[i], xn_re[i] = rfdc.sample(s)
+
+  # to convert to the converted value
+  # tofracfixed(xn_re, 16, numbits)
